requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
joblib==1.4.2
scikit-learn==1.4.2
imbalanced-learn==0.12.2
pandas==2.2.2
numpy==1.26.4
librosa==0.10.2
praat-parselmouth==0.4.3
nolds==0.5.2
pydantic==2.7.1
